feat(cobranza): F7 hito 2c-3 — los servicios dejan de escribir el espejo
Apagado del dual-write en los 4 escritores C# del censo:

- CobroService: el cobro vive SOLO como adm_pago + aplicaciones
  (transaccion_abonado_ide = null). El papel "para banco" se resuelve por
  adm_recibo_banco_pendiente (ReciboPendienteId pasa a ser su id).
- CobranzaService: el plan no escribe PLAN/PLAN-CUOTA/PLAN-PR — vive en sus
  tablas (traslado = compensación de documentos, prima = cuota mes 0).
- FacturacionMiscelaneosService: sin espejo del cargo; el misceláneo YA
  emite factura con sus líneas.
- AbonoService: el papel pendiente nace solo en su tabla.
- CaptacionPagosService (fachada): "¿ya tiene pago?" y la localización del
  pago a reversar se resuelven por adm_pago_aplicacion, no por la 201.

Tests reescritos al modelo nuevo (los que verificaban dual-write): el motor
aserta adm_pago/aplicaciones y que NO se escriben espejos; el plan aserta
cero movimientos PLAN%; los recibos pendientes viven sin fila legacy; las
fachadas verifican el documento del motor. Suite: 331 verdes.

Falta en H2c: los 4 SPs del censo (sp_lectura_v3, sp_ban_ws_pagar, NC y ND).

+++ SIAD.Core/DTOs/Cobros/CobroDtos.cs
@@ -43,8 +43,8 @@ public class CobroCrearDto
     /// <summary>Idempotencia entre canales (referencia bancaria / uuid app). Nullable en ventanilla.</summary>
     public string? ReferenciaExterna { get; set; }
 
-    /// <summary>Recibo pendiente legacy (transaccion_abonado estado 'P') que este cobro liquida.</summary>
-    public int? ReciboPendienteId { get; set; }
+    /// <summary>F7: adm_recibo_banco_pendiente.recibo_pendiente_id del papel que se está cobrando.</summary>
+    public long? ReciboPendienteId { get; set; }
 
     public DateTime? FechaPago { get; set; }
 
@@ -82,7 +82,7 @@ public class CobroResultadoDto
     public bool PolizaEncolada { get; set; }
     /// <summary>Movimiento bancario (kardex DEP) cuando el cobro fue por banco.</summary>
     public long? BanKardexId { get; set; }
-    /// <summary>ide de la fila espejo legacy (dual-write) — para recibos PDF y consultas actuales.</summary>
+    /// <summary>[OBSOLETO F7] Espejo legacy; siempre 0 desde el corte — use PagoId.</summary>
     public int TransaccionId { get; set; }
     /// <summary>True si la referencia_externa ya estaba aplicada y se devolvió el cobro original.</summary>
     public bool Idempotente { get; set; }
@@ -120,6 +120,6 @@ public class CobroDelDiaDto
     public string Estado { get; set; } = string.Empty;
     public string Usuario { get; set; } = string.Empty;
     public string? CajaNombre { get; set; }
-    /// <summary>Fila espejo legacy — para reimprimir el recibo PDF actual.</summary>
+    /// <summary>[OBSOLETO F7] Espejo legacy de cobros pre-corte; el recibo se reimprime por PagoId.</summary>
     public int? TransaccionId { get; set; }
 }
